requirements.txt
fastapi
torch
torchvision
pillow
jupyter
pandas
h5py
matplotlib
opencv-python
tqdm
